Added check if target/ dir exists

+++ scripts/publish_to_snapshots.py
@@ -267,25 +267,26 @@ class SnapshotsPublisher(object):
     def reshuffle_android_artifacts(self, src_dir):
         dst_dir = src_dir
         full_product_path = os.path.join(src_dir, product_dir_path)
-        filelist = os.listdir(full_product_path)
-        try:
-            for file in filelist:
-                src = os.path.join(full_product_path, file)
-                if os.path.isdir(src):
-                    for artifact in os.listdir(src):
-                        dest = os.path.join(dst_dir, artifact)
-                        if os.path.exists(dest):
-                            if os.path.isdir(dest):
-                                continue
-                            else:
-                                os.remove(dest)
-                        shutil.move(os.path.join(src, artifact), dest)
-        except shutil.Error:
-            print "Error while reshuffling the content"
-        try:
-            shutil.rmtree(os.path.dirname(full_product_path))
-        except shutil.Error:
-            print "Error removing empty product dir"
+        if os.path.isdir(full_product_path):
+            filelist = os.listdir(full_product_path)
+            try:
+                for file in filelist:
+                    src = os.path.join(full_product_path, file)
+                    if os.path.isdir(src):
+                        for artifact in os.listdir(src):
+                            dest = os.path.join(dst_dir, artifact)
+                            if os.path.exists(dest):
+                                if os.path.isdir(dest):
+                                    continue
+                                else:
+                                    os.remove(dest)
+                            shutil.move(os.path.join(src, artifact), dest)
+            except shutil.Error:
+                print "Error while reshuffling the content"
+            try:
+                shutil.rmtree(os.path.dirname(full_product_path))
+            except shutil.Error:
+                print "Error removing empty product dir"
 
     def move_dir_content(self, src_dir, dest_dir, sanitize=False):
         filelist = os.listdir(src_dir)
